Compile code that gets linked with libc by adding -fhosted.
The use of -fno-builtin interfers with -Wformat: no format warnings will
be given when both options are present. Because -fno-builtin means: when
you encounter a function that has the same name as a built-in function,
forget everything you know about that function.
That includes functions attributes. And without __attribute__((format...))
there will be no -Wformat warnings.

The fix is to append -fhosted when compiling something that gets linked
with libc. As command line options are processed left to right adding
-fhosted at the end overrides any earlier -fno-builtin option.

Fix compiler warnings.

Add some ugly ifdeffery to drd/drd_intercepts.c. I could not figure out how
to rewrite the offensive printf to avoid testcase breakage.

+++ helgrind/hg_intercepts.c
@@ -3550,7 +3550,7 @@ PTH_FUNC(sem_t*, semZuopen,
 
    if (TRACE_SEM_FNS) {
       fprintf(stderr, "<< sem_open(\"%s\",%ld,%lx,%lu) ",
-                      name,oflag,mode,value);
+                      name,oflag,(unsigned long)mode,value);
       fflush(stderr);
    }
 
@@ -3790,7 +3790,7 @@ static long QMutex_tryLock_WRK(void* self)
                 void *, self, long, (ret & 0xFF) ? True : False);
 
    if (TRACE_QT4_FNS) {
-      fprintf(stderr, " :: Q::tryLock -> %lu >>\n", ret);
+      fprintf(stderr, " :: Q::tryLock -> %ld >>\n", ret);
    }
    
    return ret;
@@ -3827,7 +3827,7 @@ static long QMutex_tryLock_int_WRK(void* self, long arg2)
                void *, self, long, (ret & 0xFF) ? True : False);
 
    if (TRACE_QT4_FNS) {
-      fprintf(stderr, " :: Q::tryLock(int) -> %lu >>\n", ret);
+      fprintf(stderr, " :: Q::tryLock(int) -> %ld >>\n", ret);
    }
    
    return ret;
