+++ memcheck/tests/vbit-test/irops.c
@@ -1449,6 +1449,6 @@ get_irop(IROp op)
       }
    }
 
-   fprintf(stderr, "unknown opcode %d\n", op);
+   fprintf(stderr, "unknown opcode %u\n", (unsigned)op);
    exit(1);
 }
